dbc: Add BMS battery charge and discharge power available.
Found with help of some comprehensive logs from a Kia Soul 64kWh as shared
by tiny_man on Dala's discord. They included regular UDS "read data by id"
requests sent to the BMS, which made it possible to find the values
displayed in diagnostics and then look for the same bytes in BMS messages..

ChargeAvailable sits at 180kW where SoC is high, is seen as low as 155
in a log where SocDisp is 30%.

DischargeAvailable sits at 170kW with low state of charge but comes down.
When terminating a DC fast charge with 100% SocDisp it goes as low as 50kW.
Similarly it comes up as SocDisp comes down from 100% SocDisp.

Not 100% sure these messages or the units/offsets are correct, but 95% sure
this is them.